started refactor of BlockNodeContextFactory and BlockNodeContext, is working, some tests are failing, but saving this checkpoint before further large scale refactor.
